Allowing extra information on sessions
We have now a use case where CFed sessions might not come from parent of a given commit. So we are adding a new session field to allow for non-essential fields that migth only be relevant in some cases.

So when the worker generates it, it will properly call the right function with the CFed commit