Fixing case where deleted chunks were not properly converted
We are dealing with a lot of cases where a file is deleted when it is doing session deletion. And having the bug:

```
Traceback (most recent call last):
  File "/worker/tasks/upload_processor.py", line 145, in process_async_within_lock
    **arguments,
  File "/worker/tasks/upload_processor.py", line 217, in process_individual_report
    **arguments,
  File "/worker/tasks/upload_processor.py", line 346, in do_process_individual_report
    session=session,
  File "/worker/tasks/upload_processor.py", line 366, in process_raw_upload_on_top_of_master_report
    commit_yaml, master, reports, flags, session
  File "/worker/services/report/__init__.py", line 142, in build_report_from_raw_content
    return process_raw_upload(commit_yaml, master, reports, flags, session)
  File "/worker/services/report/raw_upload_processor.py", line 70, in process_raw_upload
    sessionid, session = original_report.add_session(session or Session())
  File "/usr/local/lib/python3.7/site-packages/covreports/reports/editable.py", line 81, in add_session
    self.delete_session(sess_id_to_delete)
  File "/usr/local/lib/python3.7/site-packages/covreports/reports/editable.py", line 63, in delete_session
    file.delete_session(sessionid)
AttributeError: 'str' object has no attribute 'delete_session'
```

This happens because we expect all members of `report._chunks` at this point to be an instance of ReportFile (because we rehydrated them all). But when a ReportFile is deleted, its value in the file_summaries dictionary is deleted. So later, when recreating the Report, we missed 'rehydrating' this ReportFile because we only used the file_summaries as the source of what needs to be re-hydrated