We currently are hitting limits with a customer where they have more than 100 groups. This hits their limit on the API
So we are developing a 'make_paginated_call' functionality. This will allow for generalized listing endpoints

Due to limited time and ability to test we will only apply it to the 'list_teams' for now. But we should use it on other endpoints in the future

We might need in the future adjust the logic if Gitlab changes their pagination API considerably (they are making cursor-based pagination soon, which should not change the API, but it could)