Fixing Gitlab Enterprise to use proper config
In legacy, there was some cod that would turn configuration back into envvars (instead of the usual envvars into configuretion) that we didnt port.

It was probably done taht way to deal with the fact that torngit was open source and used by other people, so we couldnt tie to our configurations.

But now we do.