Implementing delete_session for ReportFile
The original idea is that `delete_session` was a function only in `EditableReport`s. Since they are the class used in CFF, it makes sense that they were the only situation under which we would want to delete a session (since that's what happen in CFF when a session is overwritten)

A quick look at the code will show that `EditableReportFile` is the `ReportFile` subclass used on `EditableReport`. So we thought that `EditableReport` would only have `EditableReportFile`. And therefore, `EditableReportFile` would be the
 only subclass that would need to implement `delete_session`. And that was the issue.

A `ReportFile` might show up inside an `EditableReportFile` under those conditions:

1) A flag was carriedforward (either now or in a past processing), therefore making reports be generated as an `EditableReport`
    - At the moment of creation, all members of that report are `EditableReportFile`s
2) Then, a new report is processed
    - By implementation, all member of the new report are `ReportFile`
3) The new report is merged to the existing one
    - If all filenames on the new report are already existing in the existing report, the `ReportFile`s are just merged into the `EditableReportFile` ones. No big deal
    - But if at least one filename is new, it gets appended as a `ReportFile`. There's where the problem starts
4) By itself, the above is not a problem. The report gets saved as a whole and that works fine
    - The issue appears if a second report is processed in the same task (batch) as the above report, AND if that report is deleting a previously Carriedforward Session.

If you read the above situation a couple of times, you will notice that:

An EditableReport appears, Let's say it has already 3 sessions (sessionids 0, 1 and 2). Some of them might be carriedforward. Some might not
A new Report is added to it from an upload processing, with brand new files. So it gets sessionid 3 (uploads == sessions). Let's say one of the new files is `banana_stand.py`
If you are paying attention, you see that `banana_stand.py` is a new filename that shows up in the Report. So it cannot possibly have sessionids 0, 1 or 2. If it had, it would be on the existing `EditableReport` already.
Now, in the same task (batch), the upload with sessionid 4 arrives. The system realizes it will have to overwrite a different session (or some different sessions), because it contains a flag that was originally carriedforward.
So it starts to delete that session.
Now, which of the current sessions (0, 1, 2 and 3) is definitely not carriedforward? Yes, sessionid 3.
And which file definitely only has sessionid 3? Yes, `banana_stand.py`.

So it stands to reason that when deleting sessions 0, 1 or 2, `banana_stand.py` will have to do nothing. Because it already has none of those sessions inside it.

Basically, what I am saying is that `delete_session` in the `ReportFile` should be a NO-OP function. We could add it and put a `pass` in there. Just to say it exists.

I missed doing that in the beginning.

So, just to make sure I am not forgettign some other case again, I am actually implementing the logic of `delete_session` inside `ReportFile`, but with some log.error when that happens, so we can monitor in which cases that happens.