Refactoring apply_diff
We currently have a function called `apply_diff` in the reports section of the codebase

Its purpose is to find which lines of the report are also relevant to a certain `git diff`

This function is called with a 'diff' dicionary, where that 'diff' is the 'git diff' generated by the `get_compare` functions in the `torngit` section.

But this function works a lot by having not-intuitive side-effects, such as changing some fields in the `ReportFile` level.

Our intent here is to break the function into two parts. One part that actually calculates the diff without a single side-effect on the report, and one function that has all the side-effects on the report.

This way, functions that already rely on `apply_diff` can keep using it. But functions that only want the actual calculation without the report being touched can use `calculate_diff` instead