Using new endpoint to find PRs on Gitlab
On Gitlab, the current way of finding the PRs associated with a commit is to literally list all PRs from the repository, and among those, find which repositories had a certain commit inside their 'commits' list
Not only that is very brute-force, it also did not properly paginate. But then again, pagination would make so many requests for some cases. It would be not very efficient regardless.

Gitlab has since then released an endpoint for the exact purpose of fetching the PRs associated with a specific commit.

We are then implementing the usage of this endpoint. We are also keeping the previous behavior there as a fallback for when the new one doesnt work well. We are also logging when we use taht fallback so we find out if we can eventually remove it or not