Update Load Order with Minio IAM
This change alters the ordering of the Chain provider when using iam_auth to call the IAMProviderr first in the chain. This is due to a bug in minio where fallbacks are not properly handled in the chain provider. We now have to separate auth modes for minio:

1. Using the IAMProvider if iam_auth is present
2. Using the codecov supplied minio env vars / config to connect.

The chain provider is left intact in the code in the event that minio addresses this chain problem in a future release.