Adding code so session deleting can be batched
It is not common for users to have multiple sessions being overwritten on CFF. Usually, when a client has one flag to be CFed, they are usually uploading only one session with that flag (otherwise, they risk losing too much information on overwrites, because the first upload on each commit will overwrite the 10s of uploads that were carried forward.

But it looks like some clients do it. So we will batch the logic properly