Skipping IAM test.
This commit deletes the iam_flow casette and skips the test. Since now we cannot rely on `Chain` in the underlying implementation we cannot ''trick'' minio into using the IAM auth flow, and therefore have to actually be running on an AWS instance. We can unskip this test after minio fixes their credential chain problem.