Switching Tornado for httpx
There have been a couple of connectivity problems recently in our system.

Most evidences indicate to it being a NAT gateway that is being totally used up. We already increased the per-IP limits from 128 to 4096, and from 4096 to 20K, and the problem still persists. And the fact that redeploys (which essentially kills all current instances) stops the problem for a bit makes us believe that the instances seem to be holding onto to the connection somehow until they are forcefully removed.

So we believe there is some application code not properly closing the connections, "leaking them"

We believe our combination of `tornado` + `asyncio.run` to be part of the problem because of this section https://www.tornadoweb.org/en/stable/httpclient.html#tornado.httpclient.AsyncHTTPClient.close that says

```
This method is not needed in normal use due to the way that AsyncHTTPClient objects are transparently reused. close() is generally only necessary when either the IOLoop is also being closed, or the force_instance=True argument was used when creating the AsyncHTTPClient.
```

And closing the IOLoop is exactly what asyncio.run does at the end of each run.

We don't know what kind of problem can happen because of that. We don't even know if that's the cause of our connectivity problems.

What we do know is that httpx has more explicitly closing of instances, because it works inside an `with AsyncClient()` statement. So by using it we are hoping to alleviate the connections problem a bit