Removing __enter__ from Report
The __enter__ and __exit__ from the Report are tools to enable some filtering. When entering and leaving this context, the code makes sure to erase all previously cached values, such as totals.

We don't want to miss the original values, though, and there is no clean way of storing just those in the Report. So we are trying to remove the ad-hoc filtering logic from inside the Report and move it to its own class