Removing panics from the code
Removing panics from the code and replacing them with Result<>, which is recoverable from.

Also dealing with the

```
[true, "b", [[0, true, null, null, null]]]
```

line. We have not expected `true` to be a valid coverage value. But according to

```
def line_type(line):
    """
    -1 = skipped (had coverage data, but fixed out)
    0 = hit
    1 = miss
    2 = partial
    None = ignore (because it has messages or something)
    """
    if line is True:
        return LineType.partial
    if isinstance(line, str):
        return branch_type(line)
    if line == -1:
        return LineType.skipped
    if line is False:
        return None
    if line:
        return LineType.hit
    if line is not None:
        return LineType.miss
    return None
```

true means partial. Nowhere in the code we learn what kind of partial this is (usually partials have fractions attached). So we are going with 1/2 which seems reasonable