Fixing encoding for when Github gives the wrong encoding
On some commit diff API calls Github seems to sometime tell on their response that the encoding of the response is UTF-8. But when the response is decoded, it raises an UnicodeDecodeError, clearly showing that this is not the actual response encoding.

A good example is https://api.github.com/repos/nyaruka/rapidpro/commits/30536ef28f3151846e9caab81c15ff1c6e037cbb

What it looks like is that Github is sending the contents as they are. On this example, for example, the diff content is clearly encoded in a different encoding (`latin-1`). So Github sends the response, and sets on the header

```
'content-type': 'application/vnd.github.v3.diff; charset=utf-8',
```

From my low understanding of encodings this is wrong. If Github sends content as it is and cannot attest to it being UTF-8, it should omit the charset or set to some different value.

Regardless of what Github does, we should consider, whenever an UnicodeDecodeError happen, retry decoding the content with `httpx` encoding guessing mechanism. For that we have to do something like we did: just reset the `encoding` field to `None` and let it try again


One more good indication that this is a bug and not just some misunderstanding is that both `httpx` and `vcrpy` libraries misbehave when talking to the above endpoint.

- `httpx` raises `UnicodeDecodeError` just by calling `res.text` (which is not expected behavior)
- `vcrpy` is uncapable of storing the HTTP response exactly as it is, making its cassetes useless for future testings. So we needed to use `respx` lib for mocking http calls