Slotifying ReportLine
__slots__ is good for improving memory usage and reducing attribute access time on classes that happens in big quantities

One would think that `dataclasses` is the perfect use case for using slots. But due to the way the feature was created, it's not automatic that this happens. The discussion can be found on https://github.com/ericvsmith/dataclasses/issues/28

So we are going to do it by hand

The only problem is that adding __slots__ makes it not possible to use defaults on the dataclass (explanation can also be found inside the above discussion)

Since we have so many places using the defaults, the only way is to do like below: remove the defaults and add them via an auxiliary initializer.