Repacking files
From looking at production, it seems that some chunks.txt files are getting too big because of the number of deleted files is getting big. And each deleted_file still represents a couple of lines in the final chunks.txt We need to find a way to break it up.