Letting dependencies not be tied to a version
As this is a "third-party package" from the worker/api perspectives, it is important that `shared` does not limit down the version of the packages it uses

This way those projects can update the dependencies as they need

The only one we are going to limit is minio to 6.*, because it looks like minio 7 doesn't work with us (yet).

Changing cache value of dependencies on circleci

It looks like circleCI has old packages we installed cached. We need to clean them up.

For example, we don't have pycrypto anymore. But FOSSA still finds it.