Ensuring we make no changes whenever the the 'repacking' is not needed
We want to repack a report whenever there were so many deleted files that they take most of the chunks in the archived report, which leads to a chunks file bigger than the needed.

But whenever there are no null chunk files we don't want to do the whole repacking logic. Both because we don't want to run unneeded logic and because we don't want to risk changing the report when we shouldn't (the logic already is correct, we are only being extra protective here)