Stopping flags pattern matching
As described on CE-3090, we have the current situation happening.

We found that, whenever a customer does two uploads to a commit, with the flags:

- `billing`
- `billingcustomer`

then, every time we try to calculate the totals for the `billing` flag, the uploads made with the `billingcustomer` flag would also enter the calculation (regardless of what is in the YAML). And this includes any calculation, such as per-flag statuses and comments (in the flags section).

While there are some corner cases where the user can prefer that, in general we have that this behavior is not an expected one. So we want to make it so whenever we call

```
report.filter(flags=['billing'])
```

we don't see `billingcustomer` flag coverage.