Testing fetching totals directly from the sessions
Talking about CE-2921 gave the idea of trying to use direct session totals on the flags calculations whenever the flag was only present in one session

What we are not sure of is whether the session totals are reliable enough (if they have correct values when compared to the normal calculation). We believe that yes, but want to make sure.

If this works this can help a good portion of the user notifications cases (specially the comment flags sections) even before rust comes in.

Basically, what this code is doing is "if the current filter does not filter out any paths and only matches a single session, then we can just use that session totals as the result"