present_sessions field not properly encoded on chunks
`present_sessions` is a characteristic of `EditableReportFile`s. It is created so we know which sessions are present on each ReportFile ahead of time.

But we need to also support the case where `present_sessions` contains zero elements (while still not being None).

On the specific bug the issue is that a race condition makes report.totals.lines dessynced with the chunks (more on that on CE-3153).