Changing how we calculate totals.files on filtered reports
Before anything, this change does not have any immediate effect on the user. Nowhere on statuses, comments or UI we show the number of files on a flag as a relevant metric.

When we calculate "general" totals (ie not filtered) we just take each file, calculate its totals and pass this list to the function `agg_totals`.

Among other things, that function takes the number of elements on the array to calculate its totals.files. It doesn't take into consideration anything else

This method works fine for the general method because, if a file has nothing inside it to begin with, it would not have been appended to the report. So the function can assume all files there have something.

When filtering files, something changes. If we filter by a path pattern, like `src/**`, it makes sense that files with `dist/file.js` shouldn't be counted torwards the file count. And the code already does that

But what if we are filtering by flags, and that file is in the "approved by path pattern filter" list, but has nothing on that flag? What should be the response? Is that file meant to be counted torwards the totals or not?

The current response is that yes. And that basically make every `report.filter(flags=[flag_name])` return the exact same `total.files`.

For example, if a user has on their repo a flag called `python` for the user python tests, and a flag called `javascript`, for the user js tests. There are ~100 python files and ~80 js files. None of the python tests touch anything `.js`, and vice-versa. When calculating the general totals, it makes sense to see 180. Whenever doing `report.filter(flags=["python"])`, is it less surprising for the user to see 180 or 100? Current code says 180 is less surprising. I think 100 makes more sense and is more useful.

If we look at the general example, even if a certain filename is created by the report processor, the `Report` class will not even incorporate it if it doesn't have anything inside. So an empty file without content never is counted torwards totals.files regardless of what the user wanted.

From that logic, it might make sense that the totals.files should not consider anything that, given the filter, will show no lines.

Now, about the solution: I think the best solution is to use `file_totals.lines` as a proxy for whether that file is empty or not. Right now, `file_totals.lines` is the sum of hits + partials + misses, and it makes sense that those are the only types of coverage lines to consider for this purpose. We could alternatively make a more general "separately list all the lines in this file and see if any of them has the relevant flags before calculating totals", but that is a slightly more time-costly alternative, specially if the file has 1000 lines and only the last of them has anything relevant to the flag. So I think `file_totals.lines > 0` is a good enough way to determine this.

Also adding some logging (which is where I found this discrepancy to begin with)