Limiting when ignorelines clears data
As described on a59c964a1c9c9367c23ac56cf8a9c13d8e7915bf we found that after calling `ignorelines`, the cache data on the `report._files` dict might be obsolete and nedeed clearing

We knew there was a performance deterioration for doing that. As it turns out, the time to run the UploadProcessingTask tasks increased a considerable amount.

So to try to improve this a bit, we will try to limit the cache clearing to when things actually change.

Ideally we would just remove the cache, and let the field None. The report already knows where to search for the information when the cache data is None. But for now, we don't know if anything on the API is actuvely using those fields, so we will try to be on the safe side.