Trying to fix bitbucket get_authenticated
As described before, the bitbucket previous logic was weird for private repos.

If the repo was private, the only possible response was (True, True) or the TorngitClientError

So if a repo is private and the user can see it, we assume we can edit as well

We don't want to mess with that code too much, since can_view = False might have weird consequences on the API right now. So we are just adding the proper can_edit logic to the specific section.