Logging weird rb-related behavior
Tracking some `session.totals` incompatibilities, we found one caused by this weird behavior. It seems that if the a ReportFile is ` `.rb` (ruby) and it is all misses (0% coverage), whenever a new Reportfile (with the same name) gets merged into this zero-coverage report, the misses from the old report are simply thrown away, making it look like the original upload (the one with all misses) didn't have those lines in the first place

I couldn't find any reasons other than

```
  # previous file was boil-the-ocean
  # OR previous file had END issue
```

Similar behavior when the new ReportFile has zero coverage and is `.rb`.

This particularly affects cases where the original upload has a specific flag the user wants to know about. So for example, if the first upload has a flag `unit`, by deleting a lot of those misses, the coverage for the unit flag gets affected by uploads that have nothing to do with the `unit` flag.

Tracing back to the origins of this weird code, we have:

https://github.com/codecov/codecov.io/commit/00f4244f598dd5df64adbc67c120db2fffbb5bd8

and

https://github.com/codecov/codecov.io/commit/c2a2b1037dd4ed462e08f60c3fe666d598b3fdbb

And the fact that this change was made at the same time that `include_full_missed_files` became deprecated makes me thing that this was just a way to hide full missed files that somehow managed to live until today