Using &str instead of String
It looks like that, in rust, a String is owned, and a &str can be a reference only.

pyo3 is doing the magic of automatically converting the python objects to the object types we declare at the signature of `parse_report`

From looking at https://github.com/PyO3/pyo3/blob/main/src/types/string.rs#L160-L174 I have the impression that magically converting to String includes an extra `to_owned` step. And that function to make a copy of the string so it can use a owned String object.

And:

1) It is uncessary since we are just reading the object in order to parse it.
2) For some of the chunks we read (which are 300MB in size), it might be costly

This is hunch based on no data other than the fact that the 300MB file is taking 30s to parse in rust. But I think it's worth trying.