Removing branches and partials from LineSession
In python, depite this field being there, we almost never use it.

On rust, we set it to some hardcoded values because they are not relevant to any of the calculations we were doing.

But it looks like that even this might be costly enough for rust

Locally removing those fields the performance improvement was more than 30%. We do understand that those things don't map 1:1 in alpine/musl environment. But there will probably be an improvement there at least.