Turning line into a method instead of a property
Everywhere in the code we already did lines = hits + misses + partials

But due to it being a property, it required updating the field lines every time the field hits misses or partials was updated.

This is unnecessary, and I feel this is leading to a place where someone might miss one place to update

So we are making the method get_line_count, which essentially sums up the other three values.

This way, we can remove all the updates from lines we saw before

This also leads to more consistent data