Wrapping session totals usage in function
Wrapping session totals usage in function to make it easier to determine what values are good to be used

Also adding a `CORRECT_SESSION_TOTALS_SINCE` value. This way we can add a timestamp of which sessions we want to try to use for this calculation, since we know that the totals from before 2021-04-12 deploy might be wrong.