Re-enabling parallelization
A while ago, we have parallelization enabled on this repository for some situations. Unsure what their impact was, we disabled them and then noticed they were actually deteriorating performance.

Our conclusion was that docker somehow did not play nice with that parallelization.

But now, after turning jemalloc on and reading a random comment on reddit, it looks like the original musl allocator had problems with parallelization. So we decided to retry this feature with jemalloc to test this theory.

And in fact, in our local testing, we have:

- On musl malloc, the parsing code went from 17s to ~85s (worsened 5x) when a portion of it was parallelized
- On jemalloc, the parsing code went from 6s to 1.5s (improved 4x) when the same portion of it was parallelized

So it looks like parallelization could have a positive impact now. Things on the server run behave differently than local tests, so we will still monitor it to be sure and have a baseline