Dealing with cases where coverage is higher than i64
There seems to be a couple of occasions where the value in the coverage field is a huge number. In the example it saw it was

```
18446744073709551615
```

which is `2**64 - 1`

This doesn't fit in signed i64 (signed 64 bit integers).

While this does fit in u64 (unsigned 64 bit integers), we have no guarantee the tools out there won't just pick higher numbers than this.

So we are adding the fallback as f64, a floating number field that can handle the bigger values just fine, and will be able to tell the accurate information of whether that line is a hit or miss still.

Also, we caught that in python -1 also means ignore. I don't remember seeing it causing any inconsistencies in the code. But I think we might as well get ahead of it.

Also adding some statistics to the loadfile tester