Experimenting with Cerberus validation
Experimenting with the https://docs.python-cerberus.org/en/stable/ validation library.

We currently use `voluptuous`, and we had moved from `schema` to `voluptuous` a while ago due to easier error messages.

`cerberus` is currently used on the API uploader and charts validation (for reasons I am not aware of)

The reasons for trying `cerberus` out are:

1. It's format is fairly descriptive, easier to understand, to the point where it is close to something we can put in our docs directly.
2. It's very flexible, making it possible to add validation/normalization hooks very cleanly everywhere.
3. It's currently being maintained, as opposed to `voluptuous`, which the author clearly put in "CONTRIBUTIONS ONLY" mode
4. It has a fairly clean error structure, as opposed to `schema`, which we stopped using for this very reason

The main goal here is to get to a point where we can consolidate into using one validator library only. If this works, we will also use `cerberus` for the installation-YAML validator.

We do not intend to have cerberus replace the DRF serializer, though. At least that's not part of the goal here.

We are going to implement it on the same way we put the previous library change: in shadow mode, comparing it to the main library results until we feel comfortable using it on default mode