Limitinb rb weird behavior
As documented in previous commits, there is this weird `.rb` behavior we want to remove.

While we wait for product to do the needed steps, I want to remove this behavior from "no-op" cases.

No-op cases are cases where the current totals are the same as the new file totals. Which makes deleting the previous data a no-op since the new data will have the exact same misses.

So deleting the data (which is destructive) only to have the new data be the same as the one before is objectively bad.

So we are killing this part of the behavior