Logging aggregate difference instead of difference on each step
One thing we have being doing to measure rust is logging the difference between rust and python on each run, given some minimum threshold is hit.

While this is great, this could do excessive logging, and would always miss some case like "rust improved 10ms 1000 times, to the total improvement is 10 seconds" because each diff is too small to log (this is an extreme example, not a real one).

So we will just do the aggregation inside the python level. We will have trouble with the fact that there are two reports being processed at the same time on each task, but we can figure out how to reconciliate this on the analysis script