Actually accumulate logging
Unfortunately during the previous development, we missed that the `ReadOnlyReport` keeps instantiating new `ReadOnlyReport`s on every filter call. So the `_rust_timer_so_far` values never get to reuse the previous values added up.

So we are moving this function to be part of `LazyRustReport`, which is constant across multiple filter calls