Normalizing even invalid configs
One of the features of cerberus is that is has separate steps for normalizing and validating config.

And in our case, even if the configuration is invalid, we want to normalize the config, since it's going to be used at the user yaml validation.

Also removing a print that sneaked in