Fixing branches regex on the original schema
One of the discrepancies between new validator and old was the new validator was transforming `coverage -> notify -> slack -> default -> branches` value

```
["master", "qa", "dev"]
```

into

```
["^master$", "^qa$", "^dev$"]
```

while the usual validator was not.

This usually means that the new validator is wrong

But after looking at https://github.com/codecov/codecov.io/blob/a868d063dc0aaaa7bdf9e37149ed459701c4e0ea/app/validators.py#L298 we see that the original behavior for slack (and other third party notifications) was to actually do that coercion.

So instead of fixing the new validator, we are fixing the current one

On the cases we saw logged, that usually won't change a lot (and it was only one case from one customer anyway). It will only be more strict on which branches which will get notifications.