Limiting flag names on YAML
We currently have in our docs that a flag can have a length of at most 45 chars: https://docs.codecov.com/docs/flags#one-to-one-relationship-of-flags-to-uploads

We enforce this on some of the fields. But not all of them.

Due to a discrepancy on comparing validators we found this issue. So we are updating it to also enforce at the current validator

From the logs, it looks like only one customer will have its flow affected by this, and they are not a particular heavy user.

If we do want to remove this limitation, we can consider it as a product change. But as it is, we need to enforce it at least consistently.