Replacing YAML logic
Replacing the logic using the `voluptuous` lib to use the logic from the `cerberus` lib.

This step is the actually-moving one on this migration process. For now we have been running the validator in shadow mode. And for a few weeks it hasn't shown any discrepancies when compared to the old one.

The final effect doesn't change, neither on successes or on validation errors. We even added the `error_location` logic to keep it consistent with the old one, even though the new lib now has a more complete "error_dict", which shows many errors at the same time, in a clearer dict context. The dict is also available as a property at the exception at `error_dict`.

The error messages do change. But if anything, `cerberus` makes it easier to read some of them

So we are also removing a lot of the old logic and schema generators, so we don't get confused by those.