Making configs compatible
As we consolidate config, making the api.get_config logic to use the `shared` logic, we want to make sure we don't break anything we might still need in the API

Enterprise customers `services->database_url` are the only case where the config currently seem to differ. So we are putting the API value in shared.

Also changing tests that are about minio but were checking extra stuff