Replacing panic with None
There is a condition in the `run_filereport_analysis` function that should never happen anyway, because the caller function doesn't even call it on that case.

But panic is a dangerous behavior to have. so we will just return None, which is already an accepted value for cases where "this comparison result doesn't matter"