Detecting more unexpected changes
On `get_changes_using_rust` (one of the new functions we will use for comparison analysis), we consider a change is unexpected only if `found_change["unexpected_line_changes"]`.

But a change is also unexpected if the file as a whole is unexpectedly missing or it unexpectedly appears. So we need to also consider that