Profiling Data Structure in Rust
Adding on Rust the structure to parse a profiling collection json.

This will do heavy use of the `serde_json` library. It allows one to simply create the structs with the correct types, and the library takes care of parsing the json into them (and relatively fast, the docs say). So we will create a `ProfilingDataJson` and its children properties to do that.

Then we are creating a `ProfilingData`, to be the final holder of the data and possible logic. We do this because rust says it's not doable to add `derive` properties and `pyclass` properties to the same struct. But it might end up being a positive change to split input data from actual class.

For now, we are just building a python wrapper around it with no logic. We are also adding a few tests around to check the whole thing works