Impacted Endpoints
This change implements `find_impacted_endpoints` on many levels: python and rust.

As of now, we are doing the more straightforward part of the `impacted endpoint analysys`:

For every git change, there is an removal of lines and/or an addition of lines. A changing of lines is a removal + an addition, basically.

This analysis simply looks at the removed lines, for which we have the historic profiling data for (because it already exists). If that line is present in one of the span groups, we know that that span group is affected by that line.

While that analysis is only a starter algorithm (and we will add more as we go), it still covers many cases: line changing is essentially line removal and then a corresponding line addition. So a line addition made in place of a line removal is connected to the impacted changes that are produced by that line removal.

The only case we are really not covering here are "pure" line additions, ie git changes where new code is simply added, and nothing is removed.

There are other possible analysis we can do (but are not being done on this change), but with some caveats:

- We can look at the pure line additions. Those lines are new by definition, so we can't possibly know if they themselves are run in production on a given endpoint. But what we can do is to try to analyze its neighbors.
    - For example, if lines 11-18 are being added and we know line 10 - 11 on base are run in production for a given span group (line 11 on base will become line 19 on head because of the git additions), we can say that lines 11-18 are probably impacted.
    - This is a very limited analysis, because there is not a guaranteed connection. It might be that line 10 is impacted due to function `abc`, and line 11 is impacted because of function `fgh`. The new lines added are a new function between them. This function as it is has no connection to the other ones, so there is nothing we can tell about it. And then we are in a territory where the same new code can be considered impactful depending on where the developer decides to declare it (which could be weird)
    - We can still say "those lines are _probably_ impacted because of this reason", and let the developers make best judgement. But it will take some decition

- We can also look at lines no changed by git, but changed by "unexpected changes". This one is easier than the previous points because we have a base line that we know was changes. This one is only a matter of implementing

My hopes is that by having a subset of the cases implemented, we can at least see how it will show at the PR comment, and then adding more cases later is easier.