Adding `lines_only_on_base` and `lines_only_on_head` to comparison
We already have those lines available. It was just a nicety before, but now I believe this is important information to have when putting some changes in the proper diff context.

For example, we will need to know if some lines were just removed, just added, or changed (which is removed + added) for the proper "impacted endpoint" analysis