Catching `list_repos` exception
There seems to be some cases where, on the `list_repos` Bitbucket logic, in the API call to:

```
"/repositories/%s" % (team)
```

we get a

```
{"type": "error", "error": {"message": "No workspace with identifier 'salomonvh'."}}
```

We still don't know what it is. But we believe that this indicates that doing:

```
repo["full_name"].split("/")
```

doesn't yield a "repo owner" anymore necessarily.

So we are catching this to:

1. ensure this does not block the user from syncing the other repos
2. log the necessary data so we understand what is going on