Removing skipping logging
That log is very noisy and is not that useful anymore (would be for some optimization which we will probably not be doing).

There is a similar one that is being kept. That one at least might be a guidance for some bugs and is less noisy than this one.