Fixing EOFs that were reversed
We were wrongly considering the head_file.get_eof for the base lines, and base_file.get_eof for the head lines

While this is a classic mistake, its impact was surprisingly small. It looks like it only affects case where

1. The number of total (ie. regardless of coverage) lines in the report changed between base and head
2. The last line with coverage information is different on base and head
3. The differences from 1 and 2 are on opposite directions (meaning the base file has more lines in general but the head file has the 'highest' covered line.

This seems to be relatively uncommon.

Also fixing the `get_eof` implementation to match what python does. This doesn't affect anything here (because the loops always do one "extra" check), but it is important to get it right: the eof is one line after the final covered line