Changing user yaml to something the config can set
We want to stop having the yaml secret hardcoded into the code.

So we are allowing the user to set this via config. So we mix a partly hardcoded and partly set-via-config value

We are also adding the proper field to validation so the system will not complain about it

Also taking the moment to add tests. This will encode expectations on some values that the yaml secret is supposed to work on