Extend user_schema to allow  `offset` coverage configuration. (#316)
* Extend user_schema to allow `offset` coverage configuration

Related ticket: [CODE-2161](https://codecovio.atlassian.net/browse/CODE-2161).

This allows users to select the offset coverage method to be used. Offset coverage is a way to specify what should happen to the *project* status in case a diff causes overall coverage to drop but that is part of good development practices. For that we will implement certain mechanisms to pass a project status that would otherwise fail.

Configuration will be per-status (both for all `project` statuses and custom statuses as well). This should only affect project statuses, but for the custom ones it's not a check we can make, so if `offset` is set in any patch type status it will just be ignored.

The default value will be "patch".

* Change offset field name

Following discussions and new decisions the legacy feature known as "offset coverage" was re-named in the yaml schema "removed_code_behavior". The values for it are "removals_only", "adjust_base" and "fully_covered_patch". The point of this feature is to specify a method that will allow a *project* status to pass even if coverage drops.

The 3 calculation methods are:
1. `removals_only`: pass project status if the diff consist of only lines of code being removed.
2. `adjust_base`: take the removed lines of code in the diff and remove them from the base, recalculating the coverage value for the base commit and comparing again to the head commit. That is, count the hits, misses and partials that were removed in the diff and calculate adjusted_base_coverage = (hits_base - hits_rm_diff) / ((hits_base - hits_rm_diff) + (miss_base - miss_rm_diff) + (partial_base - partial_rm_diff)), then compare adjusted_base_coverage to head_coverage.
3. `fully_covered_patch`: pass status project if patch coverage is 100%.

ps.: This config only affects *project* statuses, but it's NOT invalid to define it for *patch* status. It will just be ignored.
(I didn't find a way to make that validation)

* Update negative values for removed_code_behavior

Users will be able to cancel the use of removed_code_behavior with the 'off' string or by setting it to `False`.
This is important because the default behavior will be 'fully_covered_patch' and not 'off'