Set the stage for task_router function (#323)
Related ticket: [CODE-2328](https://codecovio.atlassian.net/browse/CODE-2328)

This is a clean-up / prep stage for the related ticket, where we'll eventaully want to put a routing function in place to route tasks based on user plan (enterprise cloud gets advantages)

First we're moving the routing of health check task to shared (previously in worker). This is because the route_task function will require a change to `task_routes` entry, and it won't be a dict anymore. That will make it harder to override in the worker (ideally won't be)

Second we're moving the billing plan names to shared. It needs knowledge of them, and doing so will prevent different names accidentally being added in api and worker.