Support webhook redelivery request (#333)
Related ticket: [CODE-2489](https://codecovio.atlassian.net/browse/CODE-2489)

These changes allow us to re-request a delivery for a webhook sent by github to our app. We'll use this to re-request installation webhooks that we missed for whatever reason, from a cron task in the worker that will run daily. This should make the codecov gh app installation process more rebust.

There were 2 options here: request the webhook delivery data directly ([this](https://codecovio.atlassian.net/browse/CODE-2489)) or do what's been done and simply ask for the webhook re-delivery. I decided to go with option 2 because the code that actually does something with the webhooks lives in the api and not in worker. So using option 1 would necessarily mean duplicating code.

Is it the best choice? I don't know, honestly. But it seems the simpler one, at least for now.