Fix the shift_lines_by_diff function (#338)
* Fix the shift_lines_by_diff function

Apparently there were cases were the `shift_lines_by_diff` function was giving wrong results.
This was caused because it was not keeping track of the difference between the current _lines index and the BASE line number (actually it was using the HEAD line number).

So first of all, if we're trying to make the coverage info from commit BASE reflect the diff and become commit HEAD we need to look to the BASE pointer, not the HEAD one. That was the first change (using index 0 of the segment header, not index 2).

Then as we process the changes (adding and deleting entries from `self._lines` the line number in BASE and that line's index in self._lines starts to diverge (it starts as -1, cause _lines is 0-indexed). So we need to carry the cumulative change from segment to segment to look in the right _lines index as we process segments.

The lines between segments don't matter because we appropriatelly add / remove entries from `self_lines` so that they go to their correct indexes if we process things correctly.

The last change was inserting `''` instead of `None`. I did that because if you create a report and append lines to it, the lines without information are filled with `''`. Just wanted to keep things consistent.

* Implementing improvement suggestions

1. Restrict the exceptions that can be caught while shifting by line.
   That is done because we might receive exceptions we don't want to catch, such as `SoftTimeLimitExceeded`.
   So we are limiting the exceptions that can be caught to the ones more likely to appear from inside the function.

2. Remove the `lines_vs_base_idx_diff`
   As pointed out by @ThiagoCodecov, the diff takes care of keeping track of that difference for us in the form of the HEAD position.
   So the assertion below is always True.
   `assert (int(segment["header"][0]) or 1) + lines_vs_base_idx_diff == (int(segment["header"][2]) or 1) - 1`

3. Improve tests by picking the old line number as coverage
   By doing that we can know exactly what happened to each line in BASE after shifting it.