Remove legacy find_pr code because of known issue (#340)
Recently we had a couple of times a situation where codecov would overwrite a PR comment with another, useless PR comment in which it would assume that the base and the head of the PR are the same commit. Worse than that it would also save the base as the head of the PR rending the UI useless.

After some investigation that you just gonna have to believe happened we tracked down the issue to the legacy find PR logic. The thing is that the PR comment includes the sha of the PR base in it. Having the sha in the comment is enough that the issues api will return it as a PR for the commit.

This situation is only a problem under the following circumstances:
1. There already is a PR comment (which includes sha for base commit B)
2. Some status is triggered for commit B and we receive a webhook because of it
3. The date the commit B was created is more recent than the head commit of the pr A.

I'll be honest in saying I'm not totally sure point 3 is a requirement.
However we had this issue twice in a week, so I guess it's valid to remove the legacy code.

This won't happen with the pulls endpoint because it only returns open PRs for commits NOT in the default branch.
For commits already merged to the default branch it only returns merged PRs, and we care little for those.