Implementing label deletion
This is the piece of logic that deletes individual labels from sessions without deleting the whole session

It loops through every line in a report, and for each line, it looks for all datapoints that have the same session id and at least some labels in common with the parameters of the function. Then it delete those datapoints

After that, it checks which sessions are still present in the datapoints array, and deletes the ones that are not there anymore (among the deleted ones, we don't touch non-deleted sessions)

If no sessions remain, we just delete the full line.