Moving CF logic one level up
We need to properly represent carryforward overwrite results to the worker processing task so it can properly update the upload objects with their states.

We also are adding partial overwrite, which is what happens on iterative testing (when only a few tests of the session are overwritten). That logic can't happen here as the overwrite needs knowledge of the fully processed report to do the overwrite, and this `add_session` has to happen before the processing, so it can pass its sessionid there. So there is no way, in the current flow, to have all the overwrite happening in this class

As we move it up, we are removing the overwrite from here. The superclass `add_session` will keep being called, as it already is.