Extend timeout for github app webhook check task. (#355)
We have seen that this task timesout constantly (every day).
Possibly the reason is that we simply have too many webhook sends to analyse.

This is one of the changes we're making. The other (https://github.com/codecov/worker/commit/c853f4209e8000e8a6b68d66a58fd4953b3366cd) is making the task run multiple times a day.
He hope to get a better understanding of what the task is doing and if it's really being as useful as we hope.