Create `SessionTotalsArray` class and refactor `reports/types.py`
We want to compress the `session_totals` arrays through the database because we're having infra problems with big customers.
The issue we have is that the position of a session in the array is used as it's index. So if we have many carryforward sessions taking up indexes, newer sessions will be thrown to the end of the array. That means we end up having a huge array full of nulls taking up space in the database. To make things worse, when you shift the report by lines we clear out the totals for some reason, so we might have just a huge empty array in the end.

The quick-and-somewhat-dirty solution we found is to encode those arrays differently. Now there's a `real_length`, with the actual size of the array (so we know the index of new sessions), and only the non-null items are stored, with their index as key in a dictionary.

That's simple enough, but I kept getting lost on all the definitions and going back and forth in the `types.py` file. That was getting in my way so I decided to separate the definitions into smaller files of related things.

Sadly this shadows my actual and important changes so this is just to save my work and refer back to it while I re-do the important changes and THEN do the refactor