Support a mixed encoding
We are still seeing some occasions where building a `Report` generates the warning "Unknown value for SessionTotal. Ignoring."
This is spread to the point that there's probably some very fundamental expectation being broken.

From logs collected recently we know the outermost data structure is a list, but the elements inside of it are not lists, as expected.
We also see consistently the session_totals being ignored are integers and "meta".
This leads me to believe we're passing a mixed format in which the new format dict is being encapsulated inside a list for some reason.

If that is the case we should see the warning go away. I added a debug line for confirmation if we want to.
The new warning should inform us of continuity of the issue.