Change base of label-analysis (#384)
Previously the base of label analysis was the HEAD of master.
The issue is that when we merge a branch into master we try to compare the head with itself.
This leads to an error and ATS doesn't run. If ATS doesn't run we can't record labels.

So I'm changing to the same value as the `worker`, which is the intersections point
between master and the current branch. Most of the time this should be the base of the
current feature branch.