fix: don't pop meta info (#22)
recently we had [issues](https://github.com/codecov/engineering-team/issues/119) with meta info missing from the encoded data
of sessions array. After much investigation and head scratching the reason is that moving data to GCS and *caching* the data
causes us to re-use the dictionary instance of files_array.

Why does this matter? because dictionaries are passed via *reference*. This means that if you change the dict reference it will
change in all pointers to that dictionary (read the *cached value* in `ArchiveField`).
So basically we were shooting ourselves in the foot.

The good news is that data is being saved and fetched correctly. We just need to not make destructive changes to the encoded data,
because it might be reused in the future.

The current changes address that, making the building of `SessionTotalArray` not destructive for the encoded data.

closes codecov/engineering-team#119