fix: suppot refresh and no refresh, change retry logic
Make sure we continue to support the current case, where github
doesn't send a refresh_token when getting the access_token

If we try to refresh wihtout the refresh_token info we return `None` and do nothing.

Also there are many endpoints from Github (under `/repos/<repo_slug>/...` that return 404
rather than 401. This is on purpose for security, but makes it harder for us.
SO we are now also retrying in 404 if the url path starts like that.

It is possible that we will try to refresh when we shouldn't.
I tried to improve the exception messages for that case, and we can revisit the retry logic
if we notice we are trying to refresh more tokens than expected and getting too many errors.