fix: fix pagination when listing PR commits
We use https://docs.github.com/en/rest/pulls/pulls?apiVersion=2022-11-28#list-commits-on-a-pull-request to list commits that belong to a PR in GitHub.
The max commits we can get is 250, per docs.
However the pagination was broken and we were getting at most 100 in `get_pull_request`
and at most 30 in `get_pull_request_commits`
(which are the max value and default value of `per_page` respectively)

These changes use pagination properly to get all the allowed 250 requests.

The fact that we know we have at most 3 pages is important because it would allow us to
fetch all 3 pages concurrently (instead of what I implemented)
Ultimately I opted for going one page at a time because realistically few commits have
more than 100 commits (I'd argue 100 commits for a PR is pretty bad already).
This means we make as few requests as possible, respecting the GH rate limit.

closes codecov/engineering-team#421