fix: make sure EditableReport is merged with EditableReportFile
When processing reports (more than 1 per `UploadProcessor` task) it's possible
that we will merge a `Report` into a `EditableReport` (the original).
This is fine, but we were not respecting the fact that files inside
`EditableReport` should be `EditableReportFile`.

These changes force all files merged into an `EditableReport` to be
of type `EditableReportFile`. Merge is left unchanged, cause it seems
complex. Instead we just run an extra check and cast to `EditableReportFile`
if needed.

closes: codecov/internal-issues#nternal-issues#138