chore: Change `labels` to `label_ids`
Changes the `CoverageDatapoint` to have `label_ids` instead of `labels`.
`labels` are preserved for now, but I want to remove them later, once the look up process is implemented.

Notice that for many functions that expected labels I simply switched.
This is because we can adjust the reports when loading them up.

Also, many of the functions are related to upload processing, and we can fix the
reports when processing (on the worker side). So it's ok to expect that no report
with labels will get to those functions.