chore: add label_index to `Report`
Part of the move with moving labels to an index table requires that the table exists.
It makes sense that it should be a part of the Report, especially during processing.
Eventually it will be saved in storage, and we can `set` it when we need to use it,
and `unset` it when we don't need it anymore to free more memory space.