Remove labels completely.
I realized that, because of the way CoverageDatapoint is encoded, and the way we create them typically
(i.e. `self.datapoints[i] = CoverageDatapoint(*cov_dp)`) the "labels" would become "label_ids" anyway.

So instead we can just remove labels completely (plus it saves space when encoded as a tuple).
To differntiate labels from indexes we make sure that indexes are turned into `int`, while labels
will remain as `str`.

I was considering if I should make the `ReportLine` fix the labels when being generated, but decided against that.
The reason being that the label_idx would not be exclusively accessed. So 2 different workers might "fix" the same
report, but for whatever reason return a different index for it. This would cause the report to be inconsistem.

Of course this can be solved by using a proper hash for the index instead of numbers, but numbers sabe a lot more space.
Instead we can trust the code that does "dangerous" operations (e.g. merge reports) to guarantee that both reports
are in a good representation of labels before doing what needs to be done.