ref: Filtered Report to not recompute lines when already called (#729) (#82)
In Filtered Report, line modifiers can take a long time to compute, that's called by lines property. We will cache the computation so that when that property is called more than once it would return the cached value. This will speedup the public reports API endpoint as it currently references lines three separate times in its serializers.