chore: Get origin repo of PRs
context: codecov/engineering-team#736

To authenticate that a tokenless upload legitimely comes from a fork
we need to check the git provider. We also need the repo where the
PR originated.

These changes also add the "slug" to the PR info, so that we can
check that the HEAD is from the fork and the BASE is the target repo
of the upload.

Notice that currently the CLI only supports tokenless for Github,
so yeah I'm only adding this to github and not other providers, but
we need to add tokenless to the others anyway.