feat: get repos from graphql (#120)
* feat: get repos from graphql

context: codecov/engineering-team#971

Adds function to get repositories from Github's GraphQL api.

It uses node_id to get the specified node from gh. This can be found in
the INSTALLATION webhook that we will process.
Along with it there will be the databaseId (service_id).

Adding the `expected_owner` entry because when processing an installation webhook
we do know what the expected owner is. Fetching the owner would require another request,
but can be done (again by node_id)

The idea is to use this function when syncing repos if we know what the repos
covered by an installation are.

* add owner lookup and pagination

There's been question if the nodes query has pagination.
So we don't trip up unecessarilly I decided to implement simple pagination ourselves.
I made page_size 100 as github docs indicate that that's the max number of records
you can get at once from endpoints that use `first` or `last` pagination.
100 repos is a lot of repos.

I was also not sure if _all_ repos will actually belong to the same owner.
Just in case I decided to add owner lookup as well. Notice it does 1 extra request
per unique owner that we encounter.
We can use this info to upsert owners if needed.