feat: make default behavior patch-centric
These changes alter the default site behavior to be default centric
_after_ `PATCH_CENTRIC_DEFAULT_TIME_START` (set to 2024-04-30).

In addition this new default is only selected for owners that have
an onboarding date (`Owner.created_at`) _after_ `PATCH_CENTRIC_DEFAULT_TIME_START`.

This is done to prevent breaking customers that might be currently relying
on the default `project` status to gatekeep PRs.

In the interest of keeping experiences consistent for an existing customer we are
also making this change per-owner (not per-repo).

context: https://github.com/codecov/engineering-team/issues/1387