feat: prevent the app from destroying itself
in a few occasions we had in staging (and 1 cloud customer) scenarios that
lead us to believe that the default app_id at times might not be present for the api pods
config.

The bad side_effect is that installations change name to 'unconfigured_app' and that's
only manually reversible. To prevent the app from destroying itself if the default
app id is not present we assume apps are configured in this scenario.