fix: make patch cuttoff date timezone aware
aka "the missing Z"

I wrongfully assumed that `fromisoformat` would generate a timezone aware datetime.
This is an issue because we generate those from the worker and api (that is, the
createstamp of owners saved in the database is pulled as being timezone aware)

And you can't compare timezone-aware and timezone-naive datetimes :clown:

So this errors out and it's pretty critical.