fix: fix fallback system for GH
There's an integration issue with the fallback system: the worker is passing the extra installation info in `data.fallback_installations` not `data.fallback_tokens`.
Personally I think that `fallback_installations` makes more sense here.

To prevent this silly (totally my fault for not testing properly locally) issue from happening again I added typehints to the `data` of TorngitBaseAdapters.

One more issue is that what's being passed as `installation_id` (that is marked as rate-limited)is the integration_id (as GitHub passes down to us), not the GithubAppInstallation instance ID.
Because it's possible to have installations from different apps we need to pass the `app_id` as well.