chore: include instance ID in GithubInstallationInfo
Up to this point we didn't pass the GithubAppInstallation ID to
the Torngit adapter. This was in part because the info was not needed,
and in part because `Owner.integration_id` doesn't have an ID.

It would be useful to have this info now, though. As it will allow
us to pin down an app to a specific app installation more easily.