fix: fix fallback system in GH (again)
I broke the GH fallback system again 😅
Fixes this issue: https://l.codecov.dev/dEsZeK

The problem is that since https://github.com/codecov/shared/pull/225
(technically https://github.com/codecov/worker/pull/470 actually) we include 1 extra piece of information in `GitHubAppInstallationInfo`, the `GitHubAppInstallation.id`.

This is breaking the function that gets tokens because we are destructuring the info dict
into the kwargs for the function (with the extra, unexpected `id` one).

To fix that we just pop it before calling the function.