feat: extend configuration for `comment.required_changes` (#259)
* feat: extend configuration for `comment.required_changes`

We are extending the configuration for required_changes.
ticket: https://github.com/codecov/engineering-team/issues/1966

I might have overengineered the different conditions, but being able to combine in both AND and OR
operations is somewhat tricky. In the end it's a satisfaction problem in that all the elements in the
conditions list need to be satisfied for the comment to occur, and the elements themselves express
and OR grouping.

* one more test

* chore: use parsing lib

Use a proper parsing lib instead of ad-hoc code to parse the required_changes
string options.
Interface remains unchanged.