feat: bundle analysis comment threshold
These changes improve documentation (and add unit tests) to `comment.require_bundle_changes`
and adds a new config key `bundle_change_threshold`.

The new key is a threshold size for the bundle in bytes. The input value
is preferably a string that accepts certain size extensions (kb, mb, gb, ...) and is
coerced into a size in bytes (as an integer)