Sync Account Users when Activating/Deactivating Users (#279)
* Add OwnerNotFound error for GitLab to appease linter

Specifically: PLE0704 Bare `raise` statement is not inside an exception handler

* Sync AccountsUsers on user activation/deactivation

This adds the feature to add and remove users from an account if the organization
that is connected to the account is activating and deactivating users.
We want to make sure that users who are deactivated are only removed when they are
not associated with _any_ other organization that the account is associated with.