feat: bundle threshold coercion
ticket: https://github.com/codecov/engineering-team/issues/2004

The bundle threshold needs to also accept percentages.
To do that we are going to re-use the percentage schema field and the
byte size schema field, and just have a wrapper around them.

A new return type has been created to make the difference between
absolute and percentage values more explicit (although technically
the values are different - float vs int - it will be easier to distinguish
if we have the type explicitly