feat: bundle status config
extends the user_schema config with the bundle_status entries.

It requires a feature flag not for any experiment, but to gate the release
of the coercion in bundle_threshold. The change is not forwards compatible,
so in the middle of deploying it's possible that old deploys would break
if the yaml was parsed by new deploys.
After a full rollout the value can be turned to True and the feature flag removed.

closes https://github.com/codecov/shared/pull/287