Fix Account User activation (#305)
There is a bug where the number of activated users count is higher than expected because it is counting _all_
the owners for all users associated with an account. Since a user may have multiple owners, it is counting
them all separately. This should now count only the number of users.