Improve tracing instrumentation
This primarily removes some excessive tracing which adds quite a bit of overhead (every line in a report used to create a tracing span).
It also adds tracing in places where it appears to be useful.

The `shared.metrics.sentry` wrapper is also being removed, as this package depends on the SDK directly, so there is no need for a shim.