fix: unpickleable exception (#325)
* fix: unpickleable exception

It seems that `celery` unpickles exceptions for some reason.
I suppose it can't do that with `InvalidInstallationError` because I was not passing all args.
Indeed the full log error is
`UnpickleableExceptionWrapper('shared.github', 'InvalidInstallationError', (), 'InvalidInstallationError()')`

Where we see it trying to create the exception with a `()` (empty tuple) for args.
Passing all args should fix this, I hope.

* add regression tests with pickle round trip