feat: add percentage change for bundles
The design for https://github.com/codecov/engineering-team/issues/1491 includes warnings in the bundle table for individual bundles.

In order to do that we need a percentage change on the bundle level.
The calculation of percentage follows the same idea as the `Comparison`: the percentage is taken using the base bundle size as the 100%