feat: add get_pipeline endpoint
related to https://github.com/codecov/engineering-team/issues/2076

Apparently there are situations in which we want to post the commit checks to more than one commit, for GitLab (when merge result pipelines are enabled).

In order to do this we have basicaly 2 options:
1. Get the extra commit SHAs, or
2. Save the extra commit SHAs.

I don't think this is such a common situation for us, so that the cost of some extra
requests to GitLab are less than the cost of making changes to the database to save the
extra info (CLI sends extra SHA and it's saved in the Upload).

So these changes introduce the endpoint that let's us get the pipeline details to see
if we need to send commit statuses to multiple SHAs.
The project ID is `Repository.service_id` and the job ID is `ReportSession.job_code`.