fix: handle better InvalidInstallationError sync
Improving on previous work that suspends installations when GitHub tell us they are suspended,
we are now also:
* Deleting installatons that can't be found;
* Cleaning up legacy integrations if they are bad.

This should reduce the number of `InvalidInstallationError` we receive from installations that we can't get a
token for.
In contrast it might increase the number of `RepositoryWithoutValidBotError`, if there are no other valid credentials
other than the invalid installation. But that is a different problem for us to fix.

closes https://github.com/codecov/internal-issues/issues/807
closes https://github.com/codecov/internal-issues/issues/809