fix: bitbucket list_repos
We are noticing non-deterministic results when syncing repos for Bitbucket.
The teams come from a `set`, so the evaluation order is not the same.

If you look closely the `page` was not being reset if we fail to get repos for a team, so the subsequent team (if any)
would start at least from page 1, and that would cause us to not find some repos.

By starting every team search in page 0 we should have that fixed.
(it actually starts at 1, but we reset the page every time)