feat: do storage service caching in shared
we currently do storage service caching in the worker with the reason
that the initialization cost is non-trivial. i'm moving this caching
to the shared layer, mostly because of the new minio rollout but also
because it generally feels like a good idea (since it's been a good idea
to do it in the worker).

There will be a change in the worker that will remove its layer of
caching.

Maybe eventually, we will find that the complication of doing this
caching is not worth it performance wise.