Add missing indices relevant for Repo Deletion
There are a few very slow `DELETE` queries as part of repo deletion.
The worst of those is deleting the newly created "shadow owner", which takes >1 minute.

Running all the `on_delete` hooks using django (which has since been fixes) has revealed that the `SET NULL` of `Commit.author` is the culprit, as that column did not have any index previously.
That is similar to the `SET NULL` on `Repository.fork`, except that query is by far not as slow, probably because of a lot less records that need to be scanned.

---

While at it, this also fixes a django warning related to the `default` not being a callable, and aligns the `Repository.fork` `SET NULL` constraint with the reality of the database.