fix: archive service bug in sentry (#521)
Previously, the archive service uses the storage service (not base
storage service) used the minio service, which can't read files
written by the new minio service, so when we enabled the new minio
feature the archive service started failing.

This change makes it so the storage service uses the new minio service
based on the value of the feature flag.