Remove unused `ReadOnlyReport.append` (#526)
* Remove unused `ReadOnlyReport.append`

This method seemed to have been intended to warn about wrong usage of `ReadOnlyReport`.

I checked both in production logs, as well as scanning through the code, and it looks like we are not misusing the this report class in such a way. So it is safe to remove this warning.

* rm test