Make some benchmark code more explicit
In particular, this better highlights the difference between a shallow and a full parse of the Report.

It also focuses the `totals` code a bit more, making the different operations more obvious.
The filtering has also been slightly updated.

As a driveby fix, this fixes the `Report.get(bind)` code, which was completely broken (because noone was using it?)