Only benchmark the `Report` class (#568)
We were previously also benchmarking the `ReadOnlyReport`. However that has yielded pretty much the same numbers, regardless of the "rust" feature flag.
I’m pretty sure that class as it exists now does not pull its weight, and I think we should remove it.

For now, I remove it from the benchmarks, as the benchmarks provide no additional value, but are just slowing down the benchmark suite.