Revert "Report caching", and remove `build_report` fn
I don’t think we have ever turned on this feature flag, as I think the Report class is not pickle-able. And either way, deserializing from `pickle` would be slower than loading it directly from storage, most likely.

For that reason, and to keep the code as simple as possible, revert this.
Along with removing the `build_report` indirection fn which only served the purpose of defaulting the `report_class`