Remove the `get_changes_using_rust` fn
This function was called within worker in "comparison mode", that ran both rust and python implementations side by side, only to always prefer the python results, and spam the logs with differences between the two that were never resolved, and I doubt anyone has looked at those in recent times.

So lets just remove this code so there is less to maintain and to worry about going forward.