feat: add metadata to MinioStorageService (#590)
GCS/S3 objects allow you to add metadata to them when writing,
and you can read that metadata back. This is done via request/response
headers. The reason I chose to pass in a metadata_container instead of
changing the return type of read_file is because I wanted to offer
the option to not care about an object's metadata, and this also makes
it so this PR won't involve a bunch of changes in worker and API to
accept 2 args as the return instead of 1.