Replace `torngit_cache` with the regular one
The `torngit_cache` was never active, as our production config has never enabled the `vcs_cache.enabled` flag.

Apart from that, the cache was used for status checks and check suites, which is something that we actively poll for changes, as it is changing as a result of CI runs.

In contrast, commits, diffs and comparisons are immutable, and a good target to cache, so lets do that.